+++ src/blocks/results/results.css
@@ -16,3 +16,7 @@
   display: none;
 }
 
+.results__container_more{
+  height: auto;
+}
+
