+++ src/index.html
@@ -67,7 +67,7 @@
       <i class="preloader__circle"></i>
     </div>
 
-    <div class="not-found">
+    <div class="not-found not-found_none">
       <h3 class="not-found__title">Ничего не найдено</h3>
       <p class="not-found__text">К сожалению по вашему запросу ничего не найдено.</p>
     </div>
